refactor(components-playground): to be built with Vite instead of Gatsby
