+++ packages/application-components/src/components/public-page-layout/public-background.js
@@ -5,3 +5,4 @@ const pngPath = path.join(__dirname, 'public-background.png');
 const base64String = fs.readFileSync(pngPath).toString('base64');
 
 module.exports = base64String;
+exports.default = base64String;
