+++ scripts/build_website.sh
@@ -7,14 +7,14 @@ echo "Preparing production builds"
 yarn compile-intl
 yarn node scripts/gatsby-cache.mjs pre
 
-echo "Building Gatsby websites"
+echo "Building documentation websites"
 
 yarn workspace @commercetools-website/custom-applications build
 yarn workspace @commercetools-website/components-playground build
 
 yarn node scripts/gatsby-cache.mjs post
 
-echo "Group websites into one folder"
+echo "Merging websites into one folder"
 
 rm -rf vercel-public
 cp -R public/website vercel-public
