+++ scripts/gatsby-cache.mjs
@@ -51,6 +51,7 @@ switch (command) {
 const workspaceRoot = findRootSync(process.cwd());
 const globalPublicPath = path.join(workspaceRoot, 'public');
 const globalCachePath = path.join(workspaceRoot, '.cache');
+const gatsbyWebsites = ['@commercetools-website/custom-applications'];
 
 const moveFolderFromTo = async (from, to) => {
   await fs.promises.mkdir(from, { recursive: true });
@@ -76,7 +77,7 @@ const run = async () => {
   await Promise.all(
     packages
       .filter((workspace) =>
-        workspace.packageJson.name.startsWith('@commercetools-website')
+        gatsbyWebsites.includes(workspace.packageJson.name)
       )
       .map(async (workspace) => {
         const namespaceKey = path.basename(workspace.dir);
