+++ visual-testing-app/package.json
@@ -35,7 +35,6 @@
   },
   "devDependencies": {
     "@vitejs/plugin-react": "2.1.0",
-    "rimraf": "3.0.2",
     "vite": "3.1.8"
   }
 }
