+++ website-components-playground/.eslintrc.js
@@ -1,0 +1,6 @@
+module.exports = {
+  rules: {
+    'react/jsx-uses-react': 'off',
+    'react/react-in-jsx-scope': 'off',
+  },
+};
