--- website-components-playground/.eslintrc.yaml
@@ -1,2 +1,0 @@
-# This file is required in order to avoid the default linter (of CRA) to throw an error
-{}
