+++ website-components-playground/index.html
@@ -12,17 +12,7 @@
       href="https://fonts.googleapis.com/css2?family=Inter:slnt,wght@-10,400;-10,500;0,400;0,500;0,600;0,700&family=Open+Sans:ital,wght@0,300;0,400;0,700;1,400;1,700&display=swap"
       rel="stylesheet"
     />
-    <title>App-Kit | Visual testing app</title>
-    <style>
-      html,
-      body {
-        margin: 0;
-        padding: 0;
-        font-size: 13px;
-        height: 100vh;
-        font-family: 'Open Sans', sans-serif;
-      }
-    </style>
+    <title>App-Kit | Components Playground</title>
   </head>
   <body>
     <div id="app"></div>
