+++ website-components-playground/src/pages/custom-form-detail-page.tsx
@@ -1,5 +1,6 @@
+import type { FormEvent } from 'react';
 import { Formik } from 'formik';
-import { FormMainPage } from '@commercetools-frontend/application-components';
+import { CustomFormDetailPage } from '@commercetools-frontend/application-components';
 import Spacings from '@commercetools-uikit/spacings';
 import Text from '@commercetools-uikit/text';
 import TextInput from '@commercetools-uikit/text-input';
@@ -31,7 +32,7 @@ const exampleCustomTitleRow = (
 
 const exampleCustomTitleRowWithTitleAndSideContent = (
   <Spacings.Inline scale="m" justifyContent="space-between">
-    <FormMainPage.PageHeaderTitle
+    <CustomFormDetailPage.PageHeaderTitle
       title="Lorem ipsum dolor sit amet, consectetur adipiscing elit."
       titleSize="big"
     />
@@ -43,23 +44,21 @@ const exampleCustomTitleRowWithTitleAndSideContent = (
   </Spacings.Inline>
 );
 
-const getCustomTitleRow = (useCustomTitleRow) => {
+const getCustomTitleRow = (useCustomTitleRow: string) => {
   switch (useCustomTitleRow) {
     case 'custom-form':
       return exampleCustomTitleRow;
     case 'custom-title-and-side-content':
       return exampleCustomTitleRowWithTitleAndSideContent;
     default:
-      break;
+      return null;
   }
 };
 
-const FormMainPageExample = (props) => {
+const CustomFormDetailPageExample = () => {
   return (
     <LayoutApp>
       <PlaygroundController
-        // eslint-disable-next-line react/prop-types
-        {...props.pageContext}
         knobs={[
           {
             kind: 'text',
@@ -90,31 +89,19 @@ const FormMainPageExample = (props) => {
             initialValue: 'default',
           },
           {
-            kind: 'text',
-            name: 'labelPrimaryButton',
-            label: 'Label Primary Button',
-            initialValue: 'Confirm',
-          },
-          {
-            kind: 'text',
-            name: 'labelSecondaryButton',
-            label: 'Label Secondary Button',
-            initialValue: 'Cancel',
-          },
-          {
             kind: 'select',
             name: 'hideControls',
             label: 'Hide Controls?',
             valueOptions: [
-              { value: false, label: 'No' },
-              { value: true, label: 'Yes' },
+              { value: 'false', label: 'No' },
+              { value: 'true', label: 'Yes' },
             ],
-            initialValue: false,
+            initialValue: 'false',
           },
         ]}
       >
         {({ values }) => (
-          <Formik
+          <Formik<{ email: string }>
             initialValues={{ email: '' }}
             validate={(formikValues) => {
               if (TextInput.isEmpty(formikValues.email)) {
@@ -125,37 +112,57 @@ const FormMainPageExample = (props) => {
             onSubmit={(formikValues) => {
               alert(`email: ${formikValues.email}`);
             }}
-            render={(formikProps) => (
-              <FormMainPage
-                title={values.title}
-                subtitle={values.subtitle}
-                customTitleRow={getCustomTitleRow(values.useCustomTitleRow)}
-                isPrimaryButtonDisabled={formikProps.isSubmitting}
-                isSecondaryButtonDisabled={formikProps.isSubmitting}
-                labelSecondaryButton={values.labelSecondaryButton}
-                labelPrimaryButton={values.labelPrimaryButton}
-                onSecondaryButtonClick={formikProps.handleReset}
-                onPrimaryButtonClick={formikProps.handleSubmit}
-                hideControls={values.hideControls}
+          >
+            {(formikProps) => (
+              <CustomFormDetailPage
+                title={values.title as string}
+                subtitle={values.subtitle as string}
+                customTitleRow={getCustomTitleRow(
+                  values.useCustomTitleRow as string
+                )}
+                onPreviousPathClick={() => alert('Go back clicked')}
+                hideControls={Boolean(values.hideControls)}
+                formControls={
+                  <>
+                    <CustomFormDetailPage.FormSecondaryButton
+                      onClick={formikProps.handleReset}
+                    />
+                    <CustomFormDetailPage.FormPrimaryButton
+                      onClick={(event) =>
+                        formikProps.handleSubmit(
+                          event as FormEvent<HTMLFormElement>
+                        )
+                      }
+                    />
+                    <CustomFormDetailPage.FormDeleteButton
+                      onClick={() => null}
+                      isDisabled={true}
+                    />
+                  </>
+                }
               >
                 <TextField
                   name="email"
                   title="Email"
                   isRequired={true}
                   value={formikProps.values.email}
-                  errors={formikProps.errors.email}
+                  errors={
+                    TextField.toFieldErrors<{ email: string }>(
+                      formikProps.errors
+                    ).email
+                  }
                   touched={formikProps.touched.email}
                   onChange={formikProps.handleChange}
                   onBlur={formikProps.handleBlur}
                 />
-              </FormMainPage>
+              </CustomFormDetailPage>
             )}
-          />
+          </Formik>
         )}
       </PlaygroundController>
     </LayoutApp>
   );
 };
-FormMainPageExample.displayName = 'FormMainPageExample';
+CustomFormDetailPageExample.displayName = 'CustomFormDetailPageExample';
 
-export default FormMainPageExample;
+export default CustomFormDetailPageExample;
