+++ website-components-playground/src/pages/form-modal-page.tsx
@@ -1,14 +1,15 @@
+import type { FormEvent } from 'react';
 import { Formik } from 'formik';
-import { CustomFormModalPage } from '@commercetools-frontend/application-components';
+import { FormModalPage } from '@commercetools-frontend/application-components';
 import TextInput from '@commercetools-uikit/text-input';
 import TextField from '@commercetools-uikit/text-field';
 import LayoutApp from '../layouts/layout-app';
 import PlaygroundController from '../components/playground-controller';
 import ModalController from '../components/modal-controller';
 
-const containerId = 'custom-form-modal-page';
+const containerId = 'form-modal-page';
 
-const CustomFormModalPageExample = () => (
+const FormModalPageExample = () => (
   <LayoutApp>
     <PlaygroundController
       knobs={[
@@ -27,25 +28,37 @@ const CustomFormModalPageExample = () => (
             'Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat.',
         },
         {
+          kind: 'text',
+          name: 'labelPrimaryButton',
+          label: 'Label Primary Button',
+          initialValue: 'Confirm',
+        },
+        {
+          kind: 'text',
+          name: 'labelSecondaryButton',
+          label: 'Label Secondary Button',
+          initialValue: 'Cancel',
+        },
+        {
           kind: 'select',
           name: 'hideControls',
           label: 'Hide Controls?',
           valueOptions: [
-            { value: false, label: 'No' },
-            { value: true, label: 'Yes' },
+            { value: 'false', label: 'No' },
+            { value: 'true', label: 'Yes' },
           ],
-          initialValue: false,
+          initialValue: 'false',
         },
       ]}
     >
       {({ values }) => (
         <ModalController
-          title="Open the Custom Form Modal Page by clicking on the button"
-          buttonLabel="Open Custom Form Modal Page"
+          title="Open the Form Modal Page by clicking on the button"
+          buttonLabel="Open Form Modal Page"
           containerId={containerId}
         >
           {({ isOpen, setIsOpen }) => (
-            <Formik
+            <Formik<{ email: string }>
               initialValues={{ email: '' }}
               validate={(formikValues) => {
                 if (TextInput.isEmpty(formikValues.email)) {
@@ -57,44 +70,45 @@ const CustomFormModalPageExample = () => (
                 alert(`email: ${formikValues.email}`);
                 setIsOpen(false);
               }}
-              render={(formikProps) => (
-                <CustomFormModalPage
-                  title={values.title}
-                  subtitle={values.subtitle}
+            >
+              {(formikProps) => (
+                <FormModalPage
+                  title={values.title as string}
+                  subtitle={values.subtitle as string}
                   isOpen={isOpen}
                   onClose={() => setIsOpen(false)}
-                  formControls={
-                    <>
-                      <CustomFormModalPage.FormSecondaryButton
-                        onClick={formikProps.handleReset}
-                      />
-                      <CustomFormModalPage.FormPrimaryButton
-                        onClick={formikProps.handleSubmit}
-                      />
-                      <CustomFormModalPage.FormDeleteButton
-                        onClick={() => null}
-                        isDisabled={true}
-                      />
-                    </>
+                  isPrimaryButtonDisabled={formikProps.isSubmitting}
+                  isSecondaryButtonDisabled={formikProps.isSubmitting}
+                  labelSecondaryButton={values.labelSecondaryButton as string}
+                  labelPrimaryButton={values.labelPrimaryButton as string}
+                  onSecondaryButtonClick={formikProps.handleReset}
+                  onPrimaryButtonClick={(event) =>
+                    formikProps.handleSubmit(
+                      event as FormEvent<HTMLFormElement>
+                    )
                   }
-                  hideControls={values.hideControls}
                   getParentSelector={() =>
-                    document.querySelector(`#${containerId}`)
+                    document.querySelector(`#${containerId}`) as HTMLElement
                   }
+                  hideControls={Boolean(values.hideControls)}
                 >
                   <TextField
                     name="email"
                     title="Email"
                     isRequired={true}
                     value={formikProps.values.email}
-                    errors={formikProps.errors.email}
+                    errors={
+                      TextField.toFieldErrors<{ email: string }>(
+                        formikProps.errors
+                      ).email
+                    }
                     touched={formikProps.touched.email}
                     onChange={formikProps.handleChange}
                     onBlur={formikProps.handleBlur}
                   />
-                </CustomFormModalPage>
+                </FormModalPage>
               )}
-            />
+            </Formik>
           )}
         </ModalController>
       )}
@@ -102,6 +116,6 @@ const CustomFormModalPageExample = () => (
   </LayoutApp>
 );
 
-CustomFormModalPageExample.displayName = 'CustomFormModalPageExample';
+FormModalPageExample.displayName = 'FormModalPageExample';
 
-export default CustomFormModalPageExample;
+export default FormModalPageExample;
