+++ website-components-playground/tsconfig.json
@@ -1,0 +1,5 @@
+{
+  "extends": "../tsconfig.json",
+  "files": ["./src/index.tsx"],
+  "include": ["./src/**/*"]
+}
